Remove onDismissRequest from filters popup
• onDismissRequest fires whenever the user clicks outside of the popup,
and this callback sets state that makes the popup close
• Clicking the filter dropdown icon should also make the popup close.
However, because of the onDismissRequest callback, clicking the dropdown
icon when the popup is visible actually does two things: 1) Close the
popup (onDismissRequest callback), 2) Open the popup (button callback).
• There does not seem to be an easy way to cause the popup to close when
the user clicks outside of the popup OR when the icon is pressed. For
now, remove the onDismissRequest callback and require the user to
manually the press the filters icon to close the popup.