Close filter options menu when selection is made
The filter options menu was changed to automatically close on filter
selection, as this UX felt more natural.

However, in doing so, another UX quirk was introduced where the new
selection checkmark would render in the menu before the menu had closed,
causing a "flashing" effect of the checkmark on filter selection.

To fix this, ListViewViewModel was modified to expose a new Flow with a
specifiable delay in milliseconds. The logic to display the checkmark
was updated to use state based on this new Flow, with just enough delay
so that the checkmark does not "jump" to the new selection before the
menu closes.

Also, filter visibility state was opportunistically moved into
ListViewViewModel to be consistent with the rest of the UI state.