Decouple UserRepository from Firebase
The User data class had implicit knowledge of Firebase via its
annotations. This meant that all implementations of UserRepository, even
mocked ones, were using a class that was tied to the Firebase
implementation under the hood. To remedy this, User is now an interface
that must be implemented for each implementation of UserRepository.