Apply client-side filtering to notes
Motivation is that Firestore does not support the type of querying
operations that we need (i.e. get all notes whose labels contain the
currently selected labels). Rather than spending time refactoring the
data model to support only a single label for notes — or worse,
reimplementing the backend to a database that supports this type of
query — it was quicker and easier to implement the desired filtering
logic on the client.