Improve navigation between top-level destinations
• All top-level destinations are now treated as "roots" — i.e. no back
stack entries are created when navigating between them. In fact, the
top-level navigation no longer uses NavHost/NavHostController for this
very reason.
• Manage child NavHostControllers at the top-level. For example, we want
to return the user to the last screen they were looking at in child
destinations with navigation, and we want to return the user to the root
of the child destination if the bottom nav icon is tapped again.