Bug: Fix memory leak possibility in validations
Implementations of the Validator interface were storing a Context as a
property, and these validators then get passed into ViewModels to
perform validations. If any ViewModel were to — intentionally or
unintentionally — hold onto a reference to one of these validators, then
the internal Context object would linger around, even after the owning
Activity/Fragment were to be destroyed (e.g. via configuration change).

The fix is to stop constructing the validation message at validation
time, and instead to construct it upfront during initialization of the
Validator instance. This way, the Context object is simply used at the
beginning but not stored, preventing the possibility of a future memory
leak.