Add and implement SessionAwareRepositoryHandler
Allows repositories to specify behavior for when the session changes.
For instance, because we now store categories in a state flow in the
categories repository, we want to be able to clear this out when the
user logs out in the (albeit rare) case that they would later log in to
a different account.