Pop backstack when going back to destination root
Originally, going to the root of the currently selected top-level
destination was implemented as a navigate() call. However, this prevents
the exitTransition animation from happening when returning to the root
in this way, as exitTransition is only shown when the backstack entry is
popped.

Also, because we are now handling the same vs different top-level
destination situations separately, the launchSingleTop flag became
unnecessary (having multiple of the same destination at the top of the
backstack is now impossible) and has therefore been removed.