Fix pop logic when navigating to selected destination root
With the previous implementation, if the global navigation root were
itself a NavGraph, then the wrong destination would be popped to because
the predicate in the last() call would match the nested NavGraph instead
of the global NavGraph.

The correct way to identify the bottom of the backstack is to check if
its startDestinationRoute is the route that we provided to
startDestination in the NavController.